+++ CHANGELOG.md
@@ -2,6 +2,9 @@
   - Put the `cabal-version` at the beginning of the generated file. This Is
     required with `cabal-version: 2.1` and higher. (see #292)
   - With `cabal-version: 2.1` or higher omit `>=` when rendering (see #292)
+  - Require `cabal-version: 2.2` when SPDX license identifiers are used (see #292)
+  - Map cabal-style licenses to SPDX license identifiers when `cabal-version`
+    is 2.2 or higher (see #292)
 
 ## Changes in 0.28.2
   - Exit with `exitFailure` on `AlreadyGeneratedByNewerHpack` or